constants, mapper, support: reorg, more scaffolding
