+++ llvm-support/Cargo.toml
@@ -14,3 +14,4 @@ version = "0.0.2-rc.2"
 [dependencies]
 paste = "1.0"
 thiserror = "1.0"
+num_enum = "0.5.3"
